+++ src/adapter/src/catalog.rs
@@ -1690,7 +1690,7 @@ impl CatalogEntry {
 struct AllocatedBuiltinSystemIds<T> {
     all_builtins: Vec<(T, GlobalId)>,
     new_builtins: Vec<(T, GlobalId)>,
-    migrated_builtins: Vec<(GlobalId, u64)>,
+    migrated_builtins: Vec<(GlobalId, String)>,
 }
 
 /// Functions can share the same name as any other catalog item type
@@ -1705,7 +1705,7 @@ pub struct SystemObjectMapping {
     object_type: CatalogItemType,
     object_name: String,
     id: GlobalId,
-    fingerprint: u64,
+    fingerprint: String,
 }
 
 pub enum CatalogItemRebuilder {
@@ -2051,8 +2051,8 @@ impl<S: Append> Catalog<S> {
                         introspection_source_index_gids
                             .get(log.name)
                             .cloned()
-                            // We don't migrate indexes so we can hardcode the fingerprint as 0
-                            .map(|id| (id, 0))
+                            // We don't migrate indexes so we can hardcode the fingerprint as ""
+                            .map(|id| (id, "".to_string()))
                     })
                     .await?;
 
@@ -2398,7 +2398,7 @@ impl<S: Append> Catalog<S> {
     /// and they need to be recreated starting at the root of the DAG and going towards the leafs.
     pub async fn generate_builtin_migration_metadata(
         &mut self,
-        migrated_ids: Vec<(GlobalId, u64)>,
+        migrated_ids: Vec<(GlobalId, String)>,
     ) -> Result<BuiltinMigrationMetadata, Error> {
         let mut migration_metadata = BuiltinMigrationMetadata::new();
 
@@ -2407,7 +2407,7 @@ impl<S: Append> Catalog<S> {
         let mut topological_sort = Vec::new();
         let mut ancestor_ids = HashMap::new();
 
-        let id_fingerprint_map: HashMap<GlobalId, u64> = migrated_ids.into_iter().collect();
+        let id_fingerprint_map: HashMap<GlobalId, String> = migrated_ids.into_iter().collect();
 
         while let Some(id) = object_queue.pop_front() {
             let entry = self.get_entry(&id);
@@ -2441,7 +2441,7 @@ impl<S: Append> Catalog<S> {
                         object_type: entry.item_type(),
                         object_name: entry.name.item.clone(),
                         id: new_id,
-                        fingerprint: *fingerprint,
+                        fingerprint: fingerprint.clone(),
                     },
                 );
             }
@@ -2577,7 +2577,7 @@ impl<S: Append> Catalog<S> {
             tx.insert_item(id, schema_id, &name, serialized_item)?;
         }
         tx.update_system_object_mappings(
-            &migration_metadata
+            migration_metadata
                 .migrated_system_table_mappings
                 .drain()
                 .collect(),
@@ -2735,7 +2735,7 @@ impl<S: Append> Catalog<S> {
     ) -> Result<AllocatedBuiltinSystemIds<T>, Error>
     where
         T: Copy + Fingerprint,
-        F: Fn(&T) -> Option<(GlobalId, u64)>,
+        F: Fn(&T) -> Option<(GlobalId, String)>,
     {
         let new_builtin_amount = builtins
             .iter()
@@ -2783,9 +2783,8 @@ impl<S: Append> Catalog<S> {
         self.storage().await.allocate_user_id().await
     }
 
-    pub async fn test_only_dont_use_in_production_allocate_system_id(
-        &mut self,
-    ) -> Result<GlobalId, Error> {
+    #[cfg(test)]
+    pub async fn allocate_system_id(&mut self) -> Result<GlobalId, Error> {
         self.storage()
             .await
             .allocate_system_ids(1)
@@ -5208,14 +5207,24 @@ impl mz_sql::catalog::CatalogItem for CatalogEntry {
 
 #[cfg(test)]
 mod tests {
+    use itertools::Itertools;
+    use std::collections::HashMap;
+    use std::error::Error;
+
+    use mz_expr::{MirRelationExpr, OptimizedMirRelationExpr};
     use mz_ore::now::NOW_ZERO;
+    use mz_repr::{GlobalId, RelationDesc, RelationType, ScalarType};
+    use mz_sql::catalog::CatalogDatabase;
     use mz_sql::names::{
         ObjectQualifiers, PartialObjectName, QualifiedObjectName, ResolvedDatabaseSpecifier,
         SchemaSpecifier,
     };
+    use mz_sql::DEFAULT_SCHEMA;
+    use mz_sql_parser::ast::Expr;
+    use mz_stash::Sqlite;
 
-    use crate::catalog::{Catalog, Op};
-    use crate::session::Session;
+    use crate::catalog::{Catalog, CatalogItem, MaterializedView, Op, Table, SYSTEM_CONN_ID};
+    use crate::session::{Session, DEFAULT_DATABASE_NAME};
 
     /// System sessions have an empty `search_path` so it's necessary to
     /// schema-qualify all referenced items.
@@ -5437,6 +5446,347 @@ mod tests {
             conn_catalog.effective_search_path(true),
             vec![mz_catalog_schema, pg_catalog_schema, mz_temp_schema]
         );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_builtin_migration() -> Result<(), Box<dyn Error>> {
+        enum ItemNamespace {
+            System,
+            User,
+        }
+
+        enum SimplifiedItem {
+            Table,
+            MaterializedView { depends_on: Vec<String> },
+        }
+
+        struct SimplifiedCatalogEntry {
+            name: String,
+            namespace: ItemNamespace,
+            item: SimplifiedItem,
+        }
+
+        impl SimplifiedCatalogEntry {
+            // A lot of the fields here aren't actually used in the test so we can fill them in with dummy
+            // values.
+            fn to_catalog_item(
+                self,
+                id_mapping: &HashMap<String, GlobalId>,
+            ) -> (String, ItemNamespace, CatalogItem) {
+                let item = match self.item {
+                    SimplifiedItem::Table => CatalogItem::Table(Table {
+                        create_sql: "TODO".to_string(),
+                        desc: RelationDesc::empty()
+                            .with_column("a", ScalarType::Int32.nullable(true))
+                            .with_key(vec![0]),
+                        defaults: vec![Expr::null(); 1],
+                        conn_id: None,
+                        depends_on: vec![],
+                    }),
+                    SimplifiedItem::MaterializedView { depends_on } => {
+                        let table_list = depends_on.iter().join(",");
+                        let depends_on = convert_name_vec_to_id_vec(depends_on, id_mapping);
+                        CatalogItem::MaterializedView(MaterializedView {
+                            create_sql: format!(
+                                "CREATE MATERIALIZED VIEW mv AS SELECT * FROM {table_list}"
+                            ),
+                            optimized_expr: OptimizedMirRelationExpr(MirRelationExpr::Constant {
+                                rows: Ok(Vec::new()),
+                                typ: RelationType {
+                                    column_types: Vec::new(),
+                                    keys: Vec::new(),
+                                },
+                            }),
+                            desc: RelationDesc::empty()
+                                .with_column("a", ScalarType::Int32.nullable(true))
+                                .with_key(vec![0]),
+                            depends_on,
+                            compute_instance: 1,
+                        })
+                    }
+                };
+                (self.name, self.namespace, item)
+            }
+        }
+
+        struct BuiltinMigrationTestCase {
+            test_name: &'static str,
+            initial_state: Vec<SimplifiedCatalogEntry>,
+            migrated_names: Vec<String>,
+            expected_previous_sink_names: Vec<String>,
+            expected_previous_materialized_view_names: Vec<String>,
+            expected_previous_source_names: Vec<String>,
+            expected_all_drop_ops: Vec<String>,
+            expected_user_drop_ops: Vec<String>,
+            expected_all_create_ops: Vec<String>,
+            expected_user_create_ops: Vec<String>,
+        }
+
+        async fn add_item(
+            catalog: &mut Catalog<Sqlite>,
+            name: String,
+            item: CatalogItem,
+            item_namespace: ItemNamespace,
+        ) -> GlobalId {
+            let id = match item_namespace {
+                ItemNamespace::User => catalog.allocate_user_id().await.unwrap(),
+                ItemNamespace::System => catalog.allocate_system_id().await.unwrap(),
+            };
+            let oid = catalog.allocate_oid().unwrap();
+            let database_id = catalog
+                .resolve_database(DEFAULT_DATABASE_NAME)
+                .unwrap()
+                .id();
+            let database_spec = ResolvedDatabaseSpecifier::Id(database_id);
+            let schema_spec = catalog
+                .resolve_schema_in_database(&database_spec, DEFAULT_SCHEMA, SYSTEM_CONN_ID)
+                .unwrap()
+                .id
+                .clone();
+            catalog
+                .transact(
+                    None,
+                    vec![Op::CreateItem {
+                        id,
+                        oid,
+                        name: QualifiedObjectName {
+                            qualifiers: ObjectQualifiers {
+                                database_spec,
+                                schema_spec,
+                            },
+                            item: name,
+                        },
+                        item,
+                    }],
+                    |_| Ok(()),
+                )
+                .await
+                .unwrap();
+            id
+        }
+
+        fn convert_name_vec_to_id_vec(
+            name_vec: Vec<String>,
+            id_lookup: &HashMap<String, GlobalId>,
+        ) -> Vec<GlobalId> {
+            name_vec.into_iter().map(|name| id_lookup[&name]).collect()
+        }
+
+        let test_cases = vec![
+            BuiltinMigrationTestCase {
+                test_name: "no_migrations",
+                initial_state: vec![SimplifiedCatalogEntry {
+                    name: "s1".to_string(),
+                    namespace: ItemNamespace::System,
+                    item: SimplifiedItem::Table,
+                }],
+                migrated_names: vec![],
+                expected_previous_sink_names: vec![],
+                expected_previous_materialized_view_names: vec![],
+                expected_previous_source_names: vec![],
+                expected_all_drop_ops: vec![],
+                expected_user_drop_ops: vec![],
+                expected_all_create_ops: vec![],
+                expected_user_create_ops: vec![],
+            },
+            BuiltinMigrationTestCase {
+                test_name: "single_migrations",
+                initial_state: vec![SimplifiedCatalogEntry {
+                    name: "s1".to_string(),
+                    namespace: ItemNamespace::System,
+                    item: SimplifiedItem::Table,
+                }],
+                migrated_names: vec!["s1".to_string()],
+                expected_previous_sink_names: vec![],
+                expected_previous_materialized_view_names: vec![],
+                expected_previous_source_names: vec!["s1".to_string()],
+                expected_all_drop_ops: vec!["s1".to_string()],
+                expected_user_drop_ops: vec![],
+                expected_all_create_ops: vec!["s1".to_string()],
+                expected_user_create_ops: vec![],
+            },
+            BuiltinMigrationTestCase {
+                test_name: "child_migrations",
+                initial_state: vec![
+                    SimplifiedCatalogEntry {
+                        name: "s1".to_string(),
+                        namespace: ItemNamespace::System,
+                        item: SimplifiedItem::Table,
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "u1".to_string(),
+                        namespace: ItemNamespace::User,
+                        item: SimplifiedItem::MaterializedView {
+                            depends_on: vec!["s1".to_string()],
+                        },
+                    },
+                ],
+                migrated_names: vec!["s1".to_string()],
+                expected_previous_sink_names: vec![],
+                expected_previous_materialized_view_names: vec!["u1".to_string()],
+                expected_previous_source_names: vec!["s1".to_string()],
+                expected_all_drop_ops: vec!["u1".to_string(), "s1".to_string()],
+                expected_user_drop_ops: vec!["u1".to_string()],
+                expected_all_create_ops: vec!["s1".to_string(), "u1".to_string()],
+                expected_user_create_ops: vec!["u1".to_string()],
+            },
+            BuiltinMigrationTestCase {
+                test_name: "multi_child_migrations",
+                initial_state: vec![
+                    SimplifiedCatalogEntry {
+                        name: "s1".to_string(),
+                        namespace: ItemNamespace::System,
+                        item: SimplifiedItem::Table,
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "u1".to_string(),
+                        namespace: ItemNamespace::User,
+                        item: SimplifiedItem::MaterializedView {
+                            depends_on: vec!["s1".to_string()],
+                        },
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "u2".to_string(),
+                        namespace: ItemNamespace::User,
+                        item: SimplifiedItem::MaterializedView {
+                            depends_on: vec!["s1".to_string()],
+                        },
+                    },
+                ],
+                migrated_names: vec!["s1".to_string()],
+                expected_previous_sink_names: vec![],
+                expected_previous_materialized_view_names: vec!["u1".to_string(), "u2".to_string()],
+                expected_previous_source_names: vec!["s1".to_string()],
+                expected_all_drop_ops: vec!["u2".to_string(), "u1".to_string(), "s1".to_string()],
+                expected_user_drop_ops: vec!["u2".to_string(), "u1".to_string()],
+                expected_all_create_ops: vec!["s1".to_string(), "u1".to_string(), "u2".to_string()],
+                expected_user_create_ops: vec!["u1".to_string(), "u2".to_string()],
+            },
+            BuiltinMigrationTestCase {
+                test_name: "topological_sort",
+                initial_state: vec![
+                    SimplifiedCatalogEntry {
+                        name: "s1".to_string(),
+                        namespace: ItemNamespace::System,
+                        item: SimplifiedItem::Table,
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "s2".to_string(),
+                        namespace: ItemNamespace::System,
+                        item: SimplifiedItem::Table,
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "u1".to_string(),
+                        namespace: ItemNamespace::User,
+                        item: SimplifiedItem::MaterializedView {
+                            depends_on: vec!["s2".to_string()],
+                        },
+                    },
+                    SimplifiedCatalogEntry {
+                        name: "u2".to_string(),
+                        namespace: ItemNamespace::User,
+                        item: SimplifiedItem::MaterializedView {
+                            depends_on: vec!["s1".to_string(), "u1".to_string()],
+                        },
+                    },
+                ],
+                migrated_names: vec!["s1".to_string(), "s2".to_string()],
+                expected_previous_sink_names: vec![],
+                expected_previous_materialized_view_names: vec!["u2".to_string(), "u1".to_string()],
+                expected_previous_source_names: vec!["s2".to_string(), "s1".to_string()],
+                expected_all_drop_ops: vec![
+                    "u1".to_string(),
+                    "u2".to_string(),
+                    "s2".to_string(),
+                    "s1".to_string(),
+                ],
+                expected_user_drop_ops: vec!["u1".to_string(), "u2".to_string()],
+                expected_all_create_ops: vec![
+                    "s1".to_string(),
+                    "s2".to_string(),
+                    "u2".to_string(),
+                    "u1".to_string(),
+                ],
+                expected_user_create_ops: vec!["u2".to_string(), "u1".to_string()],
+            },
+        ];
+
+        for test_case in test_cases {
+            let mut catalog = Catalog::open_debug_sqlite(NOW_ZERO.clone()).await?;
+
+            let mut id_mapping = HashMap::new();
+            for entry in test_case.initial_state {
+                let (name, namespace, item) = entry.to_catalog_item(&id_mapping);
+                let id = add_item(&mut catalog, name.clone(), item, namespace).await;
+                id_mapping.insert(name, id);
+            }
+
+            let migrated_ids = test_case
+                .migrated_names
+                .into_iter()
+                // We don't use the new fingerprint in this test, so we can just hard code it
+                .map(|name| (id_mapping[&name], "".to_string()))
+                .collect();
+            let migration_metadata = catalog
+                .generate_builtin_migration_metadata(migrated_ids)
+                .await?;
+
+            assert_eq!(
+                migration_metadata.previous_sink_ids,
+                convert_name_vec_to_id_vec(test_case.expected_previous_sink_names, &id_mapping),
+                "{} test failed with wrong previous sink ids",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata.previous_materialized_view_ids,
+                convert_name_vec_to_id_vec(
+                    test_case.expected_previous_materialized_view_names,
+                    &id_mapping
+                ),
+                "{} test failed with wrong previous materialized view ids",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata.previous_source_ids,
+                convert_name_vec_to_id_vec(test_case.expected_previous_source_names, &id_mapping),
+                "{} test failed with wrong previous source ids",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata.all_drop_ops,
+                convert_name_vec_to_id_vec(test_case.expected_all_drop_ops, &id_mapping),
+                "{} test failed with wrong all drop ops",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata.user_drop_ops,
+                convert_name_vec_to_id_vec(test_case.expected_user_drop_ops, &id_mapping),
+                "{} test failed with wrong user drop ops",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata
+                    .all_create_ops
+                    .into_iter()
+                    .map(|(_, _, name, _)| name.item)
+                    .collect::<Vec<_>>(),
+                test_case.expected_all_create_ops,
+                "{} test failed with wrong all create ops",
+                test_case.test_name
+            );
+            assert_eq!(
+                migration_metadata
+                    .user_create_ops
+                    .into_iter()
+                    .map(|(_, _, name)| name)
+                    .collect::<Vec<_>>(),
+                test_case.expected_user_create_ops,
+                "{} test failed with wrong user create ops",
+                test_case.test_name
+            );
+        }
 
         Ok(())
     }
