+++ src/adapter/src/catalog/builtin.rs
@@ -24,7 +24,6 @@
 
 use std::hash::Hash;
 
-use differential_dataflow::Hashable;
 use once_cell::sync::Lazy;
 use serde::Serialize;
 
@@ -139,49 +138,67 @@ pub struct BuiltinRole {
 
 /// Uniquely identifies the definition of a builtin object.
 pub trait Fingerprint {
-    fn fingerprint(&self) -> u64;
+    fn fingerprint(&self) -> String;
 }
 
-impl<T: Hash> Fingerprint for &T {
-    fn fingerprint(&self) -> u64 {
-        self.hashed()
+impl<T: TypeReference> Fingerprint for &Builtin<T> {
+    fn fingerprint(&self) -> String {
+        match self {
+            Builtin::Log(log) => log.fingerprint(),
+            Builtin::Table(table) => table.fingerprint(),
+            Builtin::View(view) => view.fingerprint(),
+            Builtin::Type(typ) => typ.fingerprint(),
+            Builtin::Func(func) => func.fingerprint(),
+            Builtin::StorageCollection(coll) => coll.fingerprint(),
+        }
     }
 }
 
 // Types and Funcs never change fingerprints so we just return constant 0
 impl<T: TypeReference> Fingerprint for &BuiltinType<T> {
-    fn fingerprint(&self) -> u64 {
-        0
+    fn fingerprint(&self) -> String {
+        "".to_string()
     }
 }
-impl Fingerprint for BuiltinFunc {
-    fn fingerprint(&self) -> u64 {
-        0
+impl Fingerprint for &BuiltinFunc {
+    fn fingerprint(&self) -> String {
+        "".to_string()
     }
 }
 
-impl<T: TypeReference> Fingerprint for &Builtin<T> {
-    fn fingerprint(&self) -> u64 {
-        match self {
-            Builtin::Log(log) => log.variant.desc().fingerprint(),
-            Builtin::Table(table) => table.desc.fingerprint(),
-            Builtin::View(view) => view.sql.hashed(),
-            Builtin::Type(typ) => typ.fingerprint(),
-            Builtin::Func(func) => func.fingerprint(),
-            Builtin::StorageCollection(coll) => coll.desc.fingerprint(),
-        }
+impl Fingerprint for &BuiltinLog {
+    fn fingerprint(&self) -> String {
+        self.variant.desc().fingerprint()
+    }
+}
+
+impl Fingerprint for &BuiltinTable {
+    fn fingerprint(&self) -> String {
+        self.desc.fingerprint()
+    }
+}
+
+impl Fingerprint for &BuiltinView {
+    fn fingerprint(&self) -> String {
+        self.sql.to_string()
+    }
+}
+
+impl Fingerprint for &BuiltinStorageCollection {
+    fn fingerprint(&self) -> String {
+        self.desc.fingerprint()
     }
 }
 
 impl Fingerprint for RelationDesc {
-    fn fingerprint(&self) -> u64 {
+    fn fingerprint(&self) -> String {
         self.typ().fingerprint()
     }
 }
 
 impl Fingerprint for RelationType {
-    fn fingerprint(&self) -> u64 {
-        self.hashed()
+    fn fingerprint(&self) -> String {
+        serde_json::to_string(self).expect("serialization cannot fail")
     }
 }
